+++ apps/frontend/translations/km.json
@@ -120,7 +120,7 @@
       "review_link": "ពិនិត្យ និងឯកភាព"
     },
     "batch_edit": {
-      "start": "កែសម្រួលទឹកជំនន់ជាក្រុម #",
+      "edit": "កែសម្រួល",
       "cancel": "បោះបង់",
       "save": "រក្សាទុក",
       "warning": "អ្នកអាចកែសម្រួលទម្រង់ជាក្រុមបានតែពីខេត្តដូចគ្នាប៉ុណ្ណោះ",
